+++ quiz-app/src/services/quiz_service.py
@@ -1,4 +1,6 @@
+import random
 from entities.quiz import Quiz
+
 
 from repositories.quiz_repository import (
     quiz_repository as default_quiz_repository
@@ -12,6 +14,13 @@ class QuestionService:
     def check_answer(self, quiz: Quiz, user_choice: int) -> bool:
         correct_answer = quiz.correct_option
         return correct_answer == user_choice
+
+    def show_questions(self):
+        questions = self._quiz_repository.get_all()
+
+        random.shuffle(questions)
+
+        return questions
 
 
 quiz_service = QuestionService()